#3883 - Scheduled routine which pulls data from ODK
git-svn-id: https://subversion.assembla.com/svn/mdss@8938 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 7e4a7f006aa926ef254c455c4acf023ff4486831 [formerly ce35ef96700de224fd62fa82b6e6ecf83a22d5e0]
Former-commit-id: 0d7c12228d67605e09983fd44bf4c98dbff4a29b

+++ trunk/src/server/dss/vector/solutions/migration/Patcher3883.java
@@ -30,7 +30,10 @@ public class Patcher3883 implements DDMSPatchIF
     {
       for (MenuItem item : it)
       {
-        item.delete();
+        if (!(item.getTerm().getTermId().equals("DDSS:0000561") || item.getTerm().getTermId().equals("MDSS:0000561")))
+        {
+          item.delete();
+        }
       }
     }
     finally
